# swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Contacts API",
        "version": "1.0.0",
        "description": "An API for managing contacts."
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local server"
        },
        {
            "url": "https://your-render-app-url.onrender.com",
            "description": "Production server"
        }
    ],
    "paths": {
        "/contacts": {
            "get": {
                "summary": "Get all contacts",
                "responses": {
                    "200": {
                        "description": "List of contacts"
                    }
                }
            },
            "post": {
                "summary": "Create a new contact",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "firstName": {
                                        "type": "string",
                                        "example": "Ada"
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "example": "Obi"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "adaobi@example.com"
                                    },
                                    "favoriteColor": {
                                        "type": "string",
                                        "example": "Blue"
                                    },
                                    "birthday": {
                                        "type": "string",
                                        "example": "1998-03-21"
                                    }
                                },
                                "required": [
                                    "firstName",
                                    "lastName",
                                    "email",
                                    "favoriteColor",
                                    "birthday"
                                ]
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Contact created successfully"
                    }
                }
            }
        },
        "/contacts/{id}": {
            "get": {
                "summary": "Get contact by ID",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Contact ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contact details"
                    },
                    "404": {
                        "description": "Contact not found"
                    }
                }
            },
            "put": {
                "summary": "Update a contact",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Contact ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "firstName": {
                                        "type": "string",
                                        "example": "Chinedu"
                                    },
                                    "lastName": {
                                        "type": "string",
                                        "example": "Okafor"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "chinedu.okafor@example.com"
                                    },
                                    "favoriteColor": {
                                        "type": "string",
                                        "example": "Green"
                                    },
                                    "birthday": {
                                        "type": "string",
                                        "example": "1995-07-19"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Contact updated successfully"
                    }
                }
            },
            "delete": {
                "summary": "Delete a contact",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "description": "Contact ID",
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contact deleted successfully"
                    },
                    "404": {
                        "description": "Contact not found"
                    }
                }
            }
        }
    }
}